third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-relative-color-expected.txt

This is a testharness.js-based test.
Found 89 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b) display-p3-linear r g b)'
  assert_true: 'color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 0 0)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 0 0)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 0 0 / 0)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 0 0 / 0)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0 g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 0 b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 0 b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 0 / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 0 / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 0)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 0)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 0 g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 0 g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 0 b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 0 b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 0 / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 0 / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 0)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 0)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0.2 g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 0.2 g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 20% g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 20% g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 0.2 b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 0.2 b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 20% b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r 20% b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 0.2 / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 0.2 / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 20% / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g 20% / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 0.2)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 0.2)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 20%)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / 20%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 0.2 g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 0.2 g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 20% g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear 20% g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 0.2 b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 0.2 b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 20% b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r 20% b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 0.2 / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 0.2 / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 20% / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g 20% / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 0.2)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 0.2)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 20%)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / 20%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 2 3 4)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 2 3 4)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 2 3 4 / 5)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 2 3 4 / 5)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -2 -3 -4)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -2 -3 -4)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -2 -3 -4 / -5)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -2 -3 -4 / -5)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 200% 300% 400%)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 200% 300% 400%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 200% 300% 400% / 500%)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear 200% 300% 400% / 500%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -200% -300% -400%)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -200% -300% -400%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -200% -300% -400% / -500%)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear -200% -300% -400% / -500%)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear g b r)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear g b r)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear b alpha r / g)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear b alpha r / g)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r r r / r)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r r r / r)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear alpha alpha alpha / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear alpha alpha alpha / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear g b r)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear g b r)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear b alpha r / g)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear b alpha r / g)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r r r / r)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r r r / r)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear alpha alpha alpha / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear alpha alpha alpha / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 1.7 1.5 1.3) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear 1.7 1.5 1.3) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 1.7 1.5 1.3) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 1.7 1.5 1.3) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 1.7 1.5 1.3 / 140%) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear 1.7 1.5 1.3 / 140%) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 1.7 1.5 1.3 / 140%) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 1.7 1.5 1.3 / 140%) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear -0.7 -0.5 -0.3) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear -0.7 -0.5 -0.3) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear -0.7 -0.5 -0.3) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear -0.7 -0.5 -0.3) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear -0.7 -0.5 -0.3 / -40%) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear -0.7 -0.5 -0.3 / -40%) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear -0.7 -0.5 -0.3 / -40%) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear -0.7 -0.5 -0.3 / -40%) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear calc(r) calc(g) calc(b))'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear calc(r) calc(g) calc(b))' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear calc(r) calc(g) calc(b) / calc(alpha))'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear calc(r) calc(g) calc(b) / calc(alpha))' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 0.8) display-p3-linear calc(r + 0.01) calc(g + 0.01) calc(b + 0.01) / calc(alpha + 0.01))'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 0.8) display-p3-linear calc(r + 0.01) calc(g + 0.01) calc(b + 0.01) / calc(alpha + 0.01))' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear calc(r * 100%) calc(g * 100%) calc(b * 100%) / calc(alpha * 100%))'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear calc(r * 100%) calc(g * 100%) calc(b * 100%) / calc(alpha * 100%))' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / calc(alpha + 0.5)) display-p3-linear r g b / calc(alpha - 0.5))'
  assert_true: 'color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / calc(alpha + 0.5)) display-p3-linear r g b / calc(alpha - 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / calc(alpha - 1.5)) display-p3-linear r g b / calc(alpha + 0.5))'
  assert_true: 'color(from color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / calc(alpha - 1.5)) display-p3-linear r g b / calc(alpha + 0.5))' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear none none none)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear none none none)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear none none none / none)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear none none none / none)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g none)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g none)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g none / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g none / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / none)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3) display-p3-linear r g b / none)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g none / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g none / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / none)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / 40%) display-p3-linear r g b / none)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear none none none) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear none none none) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear none none none / none) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear none none none / none) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 none 0.3) display-p3-linear r g b)'
  assert_true: 'color(from color(display-p3-linear 0.7 none 0.3) display-p3-linear r g b)' is a supported value for color. expected true got false
[FAIL] Property color value 'color(from color(display-p3-linear 0.7 0.5 0.3 / none) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color(display-p3-linear 0.7 0.5 0.3 / none) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property background-color value 'color(from currentColor display-p3-linear r g b)'
  assert_true: 'color(from currentColor display-p3-linear r g b)' is a supported value for background-color. expected true got false
[FAIL] Property color value 'color(from color-mix(in xyz, color(display-p3-linear 0.7 0.5 0.3), color(display-p3-linear 0.7 0.5 0.3)) display-p3-linear r g b / alpha)'
  assert_true: 'color(from color-mix(in xyz, color(display-p3-linear 0.7 0.5 0.3), color(display-p3-linear 0.7 0.5 0.3)) display-p3-linear r g b / alpha)' is a supported value for color. expected true got false
[FAIL] Property color value 'lch(from color(srgb 0.25 0.5 0.75) l c h)'
  Colors do not match.\nActual:   lch(51.4265 40.4863 262.601)\nExpected: lch(51.4321 40.4828 262.58).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 2, expected 262.58 +/- 0.02, expected 262.58 but got 262.601
[FAIL] Property color value 'color(from rgb(from color(xyz-d50 0.99 0.88 0.77) r g b) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.989872 0.879877 0.769926)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989872
[FAIL] Property color value 'color(from hsl(from color(xyz-d50 0.99 0.88 0.77) h s l) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.989872 0.879877 0.769926)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989872
[FAIL] Property color value 'color(from color(from color(xyz-d50 0.99 0.88 0.77) srgb r g b) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.989872 0.879877 0.769926)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989872
[FAIL] Property color value 'color(from color(from color(xyz-d50 0.99 0.88 0.77) display-p3 r g b) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.989831 0.879845 0.769913)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989831
[FAIL] Property color value 'color(from color(from color(xyz-d50 0.99 0.88 0.77) a98-rgb r g b) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.989614 0.879695 0.769795)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989614
[FAIL] Property color value 'color(from color(from color(xyz-d50 0.99 0.88 0.77) prophoto-rgb r g b) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.989682 0.879725 0.769775)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989682
[FAIL] Property color value 'color(from color(from color(xyz-d50 0.99 0.88 0.77) rec2020 r g b) xyz-d50 x y z)'
  Colors do not match.\nActual:   color(xyz-d50 0.98984 0.879861 0.769847)\nExpected: color(xyz-d50 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.98984
[FAIL] Property color value 'color(from rgb(from color(xyz-d65 0.99 0.88 0.77) r g b) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989888 0.879886 0.769923)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989888
[FAIL] Property color value 'color(from hsl(from color(xyz-d65 0.99 0.88 0.77) h s l) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989888 0.879886 0.769923)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989888
[FAIL] Property color value 'color(from color(from color(xyz-d65 0.99 0.88 0.77) srgb r g b) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989888 0.879886 0.769923)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989888
[FAIL] Property color value 'color(from color(from color(xyz-d65 0.99 0.88 0.77) display-p3 r g b) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989874 0.879875 0.76992)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989874
[FAIL] Property color value 'color(from color(from color(xyz-d65 0.99 0.88 0.77) a98-rgb r g b) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989656 0.879729 0.769826)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989656
[FAIL] Property color value 'color(from color(from color(xyz-d65 0.99 0.88 0.77) prophoto-rgb r g b) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989647 0.879721 0.769892)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989647
[FAIL] Property color value 'color(from color(from color(xyz-d65 0.99 0.88 0.77) rec2020 r g b) xyz-d65 x y z)'
  Colors do not match.\nActual:   color(xyz-d65 0.989824 0.879848 0.769848)\nExpected: color(xyz-d65 0.99 0.88 0.77).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.99 +/- 0.0001, expected 0.99 but got 0.989824
Harness: the test ran to completion.
